refactor(dev-server-import-maps): migrate tests from mocha/chai to node:test
Replace mocha globals, chai assertions, and hanbi spies with node:test,
node:assert/strict, and node:test mock API. Rename `test` to `test:node`.

- `import { expect } from 'chai'` -> `import assert from 'node:assert/strict'`
- `import { spy } from 'hanbi'` -> `import { mock } from 'node:test'`
- `spy()` / `spy.handler` -> `mock.fn()`
- `spy.callCount` -> `fn.mock.callCount()`
- `spy.getCall(n).args` -> `fn.mock.calls[n].arguments`
- `__dirname` -> `import.meta.dirname`
- `../src/*.js` -> `../dist/*.js`
- Add `--experimental-strip-types` to override CI NODE_OPTIONS workaround
