+++ packages/dev-server-import-maps/test/injection.test.ts
@@ -1,11 +1,12 @@
+import { it } from 'node:test';
 import { createTestServer, expectNotIncludes } from '@web/dev-server-core/test-helpers';
 import { fetchText, expectIncludes, virtualFilesPlugin } from '@web/dev-server-core/test-helpers';
 
-import { importMapsPlugin } from '../src/importMapsPlugin.js';
+import { importMapsPlugin } from '../dist/importMapsPlugin.js';
 
 it('can inject an import map into any page', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/index.html': '<html><body></body></html>',
@@ -28,7 +29,7 @@ it('can inject an import map into any page', async () => {
 
 it('can use an include pattern', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/foo/a.html': '<html><body></body></html>',
@@ -58,7 +59,7 @@ it('can use an include pattern', async () => {
 
 it('can use an exclude pattern', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/foo/a.html': '<html><body></body></html>',
@@ -86,7 +87,7 @@ it('can use an exclude pattern', async () => {
 
 it('treats directory paths with an implicit index.html file', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       {
         name: 'test',
@@ -116,7 +117,7 @@ it('treats directory paths with an implicit index.html file', async () => {
 
 it('merges with an existing import map', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/index.html':
@@ -146,7 +147,7 @@ it('merges with an existing import map', async () => {
 
 it('merges import map scopes', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/index.html':
@@ -176,7 +177,7 @@ it('merges import map scopes', async () => {
 
 it('the import map in the HTML file takes priority over the injected import map', async () => {
   const { server, host } = await createTestServer({
-    rootDir: __dirname,
+    rootDir: import.meta.dirname,
     plugins: [
       virtualFilesPlugin({
         '/index.html':
